requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
langgraph==0.0.20
langsmith>=0.0.77,<0.1.0
qdrant-client==1.7.0
streamlit==1.28.1
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
pytest==7.4.3
pytest-mock==3.12.0
euriai 